Added Node.clone; Node.height lower bound set to 0
The path-finding algorithm is always given a copy of the Grid object,
which is made in the Grid.clone method, which used to discard the height
of all the nodes (it was set to 0 in the copy).  To avoid this kind of
problems, a new Node.clone method was added, which is easier to update
to make a proper copy of the Node object, in case its definition
changes.

Also Grid.paintHall never sets a node's height to a negative value.